Refactor embeddings handlers into separate file from retrieval methods.
