+++ chainforge/rag/retrievers.py
@@ -2,175 +2,6 @@ import math, os, heapq
 from typing import List
 import numpy as np
 
-class EmbeddingModelRegistry:
-    _models = {}
-    
-    @classmethod
-    def register(cls, model_name):
-        def decorator(embedding_func):
-            cls._models[model_name] = embedding_func
-            return embedding_func
-        return decorator
-        
-    @classmethod
-    def get_embedder(cls, model_name):
-        return cls._models.get(model_name)
-    
-    @classmethod
-    def list_models(cls):
-        return list(cls._models.keys())
-
-@EmbeddingModelRegistry.register("huggingface")
-def huggingface_embedder(texts, model_name="sentence-transformers/all-mpnet-base-v2"):
-    """
-    Generate embeddings using HuggingFace Transformers.
-    
-    Args:
-        texts: List of text strings to embed
-        model_name: HuggingFace model name/path (default: sentence-transformers/all-mpnet-base-v2)
-        
-    Returns:
-        List of embeddings for each text
-    """
-    try:
-        from transformers import AutoTokenizer, AutoModel
-        import torch
-        
-        print(f"Using HuggingFace model: {model_name} for {len(texts)} texts")
-        
-        # Load model and tokenizer
-        tokenizer = AutoTokenizer.from_pretrained(model_name)
-        model = AutoModel.from_pretrained(model_name)
-        
-        embeddings = []
-        batch_size = 32
-        for i in range(0, len(texts), batch_size):
-            batch_texts = texts[i:i+batch_size]
-            batch_embeddings = []
-            
-            for t in batch_texts:
-                inputs = tokenizer(t, return_tensors="pt", truncation=True, padding=True, 
-                                  max_length=512)  # Add max_length for safety
-                with torch.no_grad():
-                    outputs = model(**inputs)
-                # Use mean pooling by default
-                emb = outputs.last_hidden_state.mean(dim=1).squeeze().tolist()
-                batch_embeddings.append(emb)
-                
-            embeddings.extend(batch_embeddings)
-            
-        return embeddings
-    except Exception as e:
-        print(f"HuggingFace embedder failed: {str(e)}")
-        raise ValueError(f"Failed to generate HuggingFace embeddings: {str(e)}")
-
-@EmbeddingModelRegistry.register("OpenAI Embeddings")
-def openai_embedder(texts, model_name="text-embedding-ada-002"):
-    """
-    Generate embeddings using OpenAI Embeddings.
-    
-    Args:
-        texts: List of text strings to embed
-        model_name: OpenAI embedding model to use (default: text-embedding-ada-002)
-        
-    Returns:
-        List of embeddings for each text
-    """
-    try:
-        import openai
-        print(f"Using OpenAI model: {model_name} for {len(texts)} texts")
-        
-        embeddings = []
-        # Process in batches of 16 to stay within rate limits
-        batch_size = 16
-        for i in range(0, len(texts), batch_size):
-            batch_texts = texts[i:i+batch_size]
-            batch_embeddings = []
-            
-            for t in batch_texts:
-                resp = openai.Embedding.create(input=t, model=model_name)
-                emb = resp["data"][0]["embedding"]
-                batch_embeddings.append(emb)
-                
-            embeddings.extend(batch_embeddings)
-            
-        return embeddings
-    except Exception as e:
-        print(f"OpenAI embedder failed: {str(e)}")
-        raise ValueError(f"Failed to generate OpenAI embeddings: {str(e)}")
-
-@EmbeddingModelRegistry.register("Cohere Embeddings")
-def cohere_embedder(texts, model_name="embed-english-v2.0"):
-    """
-    Generate embeddings using Cohere Embeddings.
-    
-    Args:
-        texts: List of text strings to embed
-        model_name: Cohere embedding model to use (default: embed-english-v2.0)
-        
-    Returns:
-        List of embeddings for each text
-    """
-    try:
-        import cohere
-        print(f"Using Cohere model: {model_name} for {len(texts)} texts")
-        
-        # Get API key from environment or settings
-        api_key = os.environ.get("COHERE_API_KEY")
-        if not api_key:
-            from flask import current_app
-            api_key = current_app.config.get("COHERE_API_KEY")
-            
-        if not api_key:
-            raise ValueError("Cohere API key not found in environment or app config")
-            
-        co = cohere.Client(api_key)
-        
-        batch_size = 32  
-        embeddings = []
-        
-        for i in range(0, len(texts), batch_size):
-            batch_texts = texts[i:i+batch_size]
-            response = co.embed(texts=batch_texts, model=model_name)
-            embeddings.extend(response.embeddings)
-            
-        return embeddings
-    except Exception as e:
-        print(f"Cohere embedder failed: {str(e)}")
-        raise ValueError(f"Failed to generate Cohere embeddings: {str(e)}")
-
-@EmbeddingModelRegistry.register("Sentence Transformers")
-def sentence_transformers_embedder(texts, model_name="all-MiniLM-L6-v2"):
-    """
-    Generate embeddings using Sentence Transformers.
-    
-    Args:
-        texts: List of text strings to embed
-        model_name: Sentence Transformers model name (default: all-MiniLM-L6-v2)
-        
-    Returns:
-        List of embeddings for each text
-    """
-    try:
-        from sentence_transformers import SentenceTransformer
-        print(f"Using SentenceTransformer model: {model_name} for {len(texts)} texts")
-        
-        model = SentenceTransformer(model_name)
-        
-        # Process in reasonable batch sizes
-        batch_size = 32
-        embeddings = []
-        
-        for i in range(0, len(texts), batch_size):
-            batch_texts = texts[i:i+batch_size]
-            embeddings = model.encode(batch_texts).tolist()
-            embeddings.extend(embeddings)
-            
-        return embeddings
-    except Exception as e:
-        print(f"SentenceTransformer embedder failed: {str(e)}")
-        raise ValueError(f"Failed to generate SentenceTransformer embeddings: {str(e)}")
-
 # Define a registry for retrieval methods
 class RetrievalMethodRegistry:
     _methods = {}
